Fix fifteen findings from the review of phases 8 and 9
Four agents reviewed the memory and telemetry phases: the domains, the shared
plumbing, and the tests. What they found, most serious first.

Postgres did not actually work, which is the thing phase 9 claimed to deliver:

- telephony_events.mono_ns and the two span clocks were INTEGER, four bytes on
  Postgres. time.monotonic_ns() outgrows that 2.15 seconds after boot, so every
  event and span write was refused -- and the recorder swallows a failed batch,
  so the dashboard looked populated with an empty timeline. New BigInt type,
  migration 0016, and a test that writes a real clock reading.
- store.fetch() bound its parameters untyped, so every windowed read -- the
  overview, call search, every alert -- sent an ISO string at a TIMESTAMP
  column and 500'd. Timestamps also came back as naive datetimes rather than
  the ISO strings the dashboard parses, which is a silent timezone shift.
  Both fixed, and queries.* now runs against Postgres in the suite.
- Memory search pasted the user's words into to_tsquery, which parses them as
  syntax: "Hey! where is the boiler?" raised instead of searching, and
  core.agent swallows it, so recall quietly stopped working for ordinary
  speech. websearch_to_tsquery never raises. The FTS5 side now escapes a quote
  rather than ending its string on one.

The pooled SQLite connection is gone, and with it a bug of my own making:

- Phase 9 put telemetry on one pooled connection, which holds the file's write
  lock between units of work. It failed one driven call in four with "database
  is locked" -- starving the other writers that share the file.
- Measured, the pool bought nothing: audio-loop p95 is 0.3-1.0 ms without it
  against 0.9-4.7 ms with it. Only the off-path drain got slower (660 ms vs
  230 ms), which is explicitly allowed to be. The full suite also went from
  33 minutes to 4, having been serialised on that lock all along.
- So `pooled` is removed from get_engine, _create_engine, session_scope and
  DatabaseService rather than tuned. That also retires three findings about
  it: a duplicate engine on Postgres, a cache key built by string
  concatenation that a database path could alias, and its untested pragmas.

Startup and configuration:

- Two processes migrating one fresh SQLite file corrupted each other -- no
  transactional DDL, so the loser was not rolled back. The prod compose file
  starts pincer and pincer-tasks on one volume, so a first deploy hit exactly
  this. Serialised by a lock file; the comment claiming it was harmless was
  wrong.
- PINCER_DATABASE_URL=sqlite+aiosqlite://... killed every entry point with
  MissingGreenlet. get_sync_url now normalises the driver for both backends.

Memory:

- full_text applied the tag filter after the search engine's LIMIT, dropping a
  tagged match ranked below it; and it read the caller's whole memories table,
  embeddings included, on every turn (872 ms and 158 MB at 20k memories). The
  filter is pushed into the search SQL and only the ranked rows are read.
- store_entity could hand back an id for a row that does not exist;
  upsert_entity returns the id that won, and get_entities is unbounded again.

Tests that looked like they covered something and did not. Each of these was
confirmed by breaking the code and watching the suite stay green:

- the upsert rules were still duplicated in store.py as dead code, under the
  comment explaining why they matter -- a maintainer adding a counter there
  would change nothing
- two asserts on values the test itself supplied
- ranking asserted as a set, so a reversed sort passed (core.agent keeps the
  top three, so that is the least relevant memories, every turn)
- the with_embedding filter, whose loss is a crash on the first similarity
  search
- get_sync_url asserted with a password-less URL, so redaction would not show
- _named had no test at all, including its mismatch branch
- init_database's db_path=None branch and its return value
- the stray-sqlite guard missed import_module("aiosqlite") and sqlite3
  entirely; it is an AST scan now
- the Postgres db_url fixture handed every test the same database, so two runs
  against one server would pull each other's tables

Every test added here was verified to fail without its fix.

Known and not fixed: tests/voice_harness/test_receptionist.py::
test_after_hours_message fails about a quarter of runs. Bisected to 0a29cca
(phase 8): 0/10 on its parent, 2-4/10 from there on, unchanged by this commit.
Raising busy_timeout to 30 s turns a 25 s failure into a 253 s one, so a
connection holds the write lock indefinitely rather than slowly -- a leaked
transaction, not contention. It disappears under instrumentation.

+++ src/pincer/models/telephony.py
@@ -4,7 +4,7 @@ import sqlalchemy as sa
 from sqlalchemy import Column, Index, Integer, Text
 from sqlmodel import Field, SQLModel
 
-from pincer.db.types import IsoText, Real
+from pincer.db.types import BigInt, IsoText, Real
 
 
 class TelephonyCall(SQLModel, table=True):
@@ -75,7 +75,7 @@ class TelephonyEvent(SQLModel, table=True):
     turn_id: str | None = Field(default=None, sa_column=Column(Text(), server_default=sa.text("''")))
     name: str = Field(sa_column=Column(Text(), nullable=False))
     ts_utc: str = Field(sa_column=Column(IsoText(), nullable=False))
-    mono_ns: int | None = Field(default=None, sa_column=Column(Integer(), server_default=sa.text("0")))
+    mono_ns: int | None = Field(default=None, sa_column=Column(BigInt(), server_default=sa.text("0")))
     seq: int | None = Field(default=None, sa_column=Column(Integer(), server_default=sa.text("0")))
     attributes: str | None = Field(default=None, sa_column=Column(Text(), server_default=sa.text("'{}'")))
 
@@ -95,8 +95,8 @@ class TelephonySpan(SQLModel, table=True):
     name: str = Field(sa_column=Column(Text(), nullable=False))
     start_utc: str = Field(sa_column=Column(IsoText(), nullable=False))
     end_utc: str | None = Field(default=None, sa_column=Column(IsoText()))
-    start_mono_ns: int | None = Field(default=None, sa_column=Column(Integer(), server_default=sa.text("0")))
-    end_mono_ns: int | None = Field(default=None, sa_column=Column(Integer()))
+    start_mono_ns: int | None = Field(default=None, sa_column=Column(BigInt(), server_default=sa.text("0")))
+    end_mono_ns: int | None = Field(default=None, sa_column=Column(BigInt()))
     duration_ms: float | None = Field(default=None, sa_column=Column(Real()))
     status: str | None = Field(default=None, sa_column=Column(Text(), server_default=sa.text("'ok'")))
     attempt: int | None = Field(default=None, sa_column=Column(Integer(), server_default=sa.text("1")))
